// src/apis/models/RiskLevelRuleActions.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RuleAction } from './RuleAction';
import { HttpFile } from '../http/http';

export class RiskLevelRuleActions {
  'VERY_HIGH': RuleAction;
  'HIGH': RuleAction;
  'MEDIUM': RuleAction;
  'LOW': RuleAction;
  'VERY_LOW': RuleAction;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'VERY_HIGH',
      baseName: 'VERY_HIGH',
      type: 'RuleAction',
      format: '',
    },
    {
      name: 'HIGH',
      baseName: 'HIGH',
      type: 'RuleAction',
      format: '',
    },
    {
      name: 'MEDIUM',
      baseName: 'MEDIUM',
      type: 'RuleAction',
      format: '',
    },
    {
      name: 'LOW',
      baseName: 'LOW',
      type: 'RuleAction',
      format: '',
    },
    {
      name: 'VERY_LOW',
      baseName: 'VERY_LOW',
      type: 'RuleAction',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return RiskLevelRuleActions.attributeTypeMap;
  }

  public constructor() {}
}